agenda mesa updating
