+++ src/app/shared/boton/boton-descargar/boton-descargar.component.html
@@ -1,0 +1,3 @@
+<button nz-button nzType="link" nz-tooltip nzTooltipTitle="Descargar" (click)="descargarPdf(rutaArchivo)">
+    <span nz-icon nzType="file-excel" nzTheme="outline"></span>
+</button>
