+++ src/app/shared/shared.module.ts
@@ -4,6 +4,8 @@ import { GeoChartComponent } from './geo-chart/geo-chart.component';
 import { TableCardComponent } from './table-card/table-card.component';
 import { SliderTinyComponent } from './slider-tiny/slider-tiny.component';
 import { EstadoTagComponent } from './estado-tag/estado-tag.component';
+import { BotonDescargarComponent } from './boton/boton-descargar/boton-descargar.component';
+
 
 
 
@@ -14,14 +16,16 @@ import { EstadoTagComponent } from './estado-tag/estado-tag.component';
     GeoChartComponent,
     TableCardComponent,
     SliderTinyComponent,
-    EstadoTagComponent
+    EstadoTagComponent,
+    BotonDescargarComponent
   ],
   exports: [
     ChartComponent,
     GeoChartComponent,
     TableCardComponent,
     SliderTinyComponent,
-    EstadoTagComponent
+    EstadoTagComponent,
+    BotonDescargarComponent
   ]
 })
 export class SharedModule { }
